server/src/requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
